Double sending + message sending indicator
Added a Cupertino progress view to indicate when a multimedia message is sending (since that can take awhile). Had to use a Cupertino progress view because Snackbar and Toast didn't work. Also added code to prevent messages from double sending by blocking the send function if sending is in progress.