ViewPersonDetails interactivity enabled
Additionally, I realized that I can animate the search bar in and out by using a SizeTransition in a custom builder inside AnimatedSwitcher, which is what I did in MyProfileTab (replaced the slide transition with a size transition because it looks a lot better). Also increased the column size from 25 to 35 to make it easier to tap the delete button to remove extra images.