You can now create a pod directly from the AddPersonToPod view
The big issue was apparently I had to call setState in the .then() block after navigating in order to the view to update with the new pod. Very strange, but that's how it is. Also, in most cases, deleting a pod will automatically navigate you back from the pod screen. Even in cases where that fails, the Join button is removed so nothing bad can happen.