Pod images and document now delete successfully if creation cancelled
I call the cleanup function inside a dispose() block (to handle cases where the user goes back) and in a onForegroundLost block (to handle cases where the user closes the app). I don't call the function in onFocusLost() or onVisibilityLost() because that can cause data to get deleted when the user opens a text field, which is not desired.