Solved a MessagingViewI Issue
The problem was that I was calling .first and .last in updateAnimatedList(), but I forgot that .first and .last aren't null-safe. So if I tried sending a message after deleting the conversation, the new list wouldn't update because of the crash. That has been fixed with some manual null checks.