Reorganized the WelcomeView tab view
The reason is that in Flutter (unlike iOS), running Bluetooth with Google Nearby does NOT slow down internet connection. This is a good thing, meaning that Discover Nearby can run continuously (instead of having to stop whenever the user left the screen to ensure internet speeds were acceptable). Since there's no need for a Discover Nearby screen in Flutter, I simply took it out and added My Pods to the main tab view.