Marking messages as read now works
As a result, bolding/unbolding the LatestMessageRow conversation preview works. I also moved the startPublishAndSubscribe() command inside the FutureBuilder in Main.dart to ensure that my profile data is available before I start scanning for nearby users. Finally, I added Bluetooth permissions to the Android manifest; that would probably be necessary in order for the app to get approved on the Google Play store.