Moved splash screen image to /res/drawable-hdpi and updated build version from 2 to 3
android_splash_screen_smaller moved from res/drawable to res/drawable-hdpi to stop the app from crashing on startup on certain devices due to a rendering error. Also changed the build code to version 1.0.0+3 and added an if (mounted) block in front of setState() in PeopleIMetBackendFunctions.shared.sortedListOfPeople.addListener() to stop the app from throwing silent errors if a person is met before the app state can be initialized (not that important, but just to be safe).