Fixed the ages not showing in list row bug
Fixed the bug that caused ages not to show in pod members or people I met except for myself (I simply forgot to pass ages into the list row object). Also fixed a couple typos and a couple .get() calls to document objects that would cause silent failures in the background if the field didn't exist. Also changed version code to 5.