Fixed the bug involving empty Bio's
There was a bug where if a user didn't write a bio, their bio would say "Bio" instead of being empty. This had to do with errors in handling empty strings in the client-side code, which has since been fixed. I also fixed a bug where the pod name wouldn't show up in the warning when deleting a pod message and changed it so that when deleting a DM, it will only say "this message will be deleted for [FIRST NAME] as well" instead of "this message will be deleted for [FULL NAME] as well". Using the first name only helps make the alert shorter and just seems more natural.